# server/internal/docs/openapi.yaml
openapi: 3.0.3
info:
    title: ural-hackaton API
    version: 1.0.0
    description: API documentation for the ural-hackaton project.
servers:
    - url: http://localhost:8080
      description: Local development server
...
paths:
...
    /requests:
        post:
            tags: [Requests]
            summary: Create request
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateRequestDto"
            responses:
                "201":
                    description: Request created
    /requests/{id}:
        get:
            tags: [Requests]
            summary: Get request by id
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    description: Request found
                "404":
                    description: Request not found
    /requests/user/{user_id}:
        get:
            tags: [Requests]
            summary: Get requests by user id
            parameters:
                - in: path
                  name: user_id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    description: Request list
    /bookings:
        get:
            tags: [Bookings]
            summary: Get all bookings
            responses:
                "200":
                    description: Booking list
        post:
            tags: [Bookings]
            summary: Create booking
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/CreateBookingDto"
            responses:
                "201":
                    description: Booking created
        put:
            tags: [Bookings]
            summary: Update booking
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Booking"
            responses:
                "200":
                    description: Booking updated
    /bookings/{id}:
        get:
            tags: [Bookings]
            summary: Get booking by id
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    description: Booking found
                "404":
                    description: Booking not found
        delete:
            tags: [Bookings]
            summary: Delete booking
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                "204":
                    description: Deleted
    /bookings/user/{user_id}:
        get:
            tags: [Bookings]
            summary: Get bookings by user id
            parameters:
                - in: path
                  name: user_id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                "200":
                    description: Booking list
components:
    schemas:
...
        CreateRequestDto:
            type: object
            required: [request_message, user_id]
            properties:
                request_message:
                    type: string
                user_id:
                    type: integer
                    format: int64
...
        CreateBookingDto:
            type: object
            required: [booking_date, booking_zone, booking_slots, user_id]
            properties:
                booking_date:
                    type: string
                    format: date-time
                booking_zone:
                    type: string
                booking_slots:
                    type: integer
                user_id:
                    type: integer
                    format: int64
        Booking:
            type: object
            properties:
                booking_id:
                    type: integer
                    format: int64
                booking_date:
                    type: string
                    format: date-time
                booking_zone:
                    type: string
                booking_slots:
                    type: integer
                user_id:
                    type: integer
                    format: int64